refactoring and cleaning VolumeImageViewer by using new channel and bledingInformation classes
